Creating the controlller, DTO and
After we create the entities, we set the DTO(class with expected data type) and make the verifications with the useCase. So, we use the execute function to verify if the student can make the test. After we get the response(0 or 1), the controller returns the message showing if the student can make or not the test.